Refactor catalog eval lookup logic
Resolves #780
